* src/pam_krb5_storetmp.c: silence a compiler warning comparing the 	return code from write() with strlen().

+++ src/pam_krb5_storetmp.c
@@ -151,7 +151,7 @@ main(int argc, const char **argv)
 success:
 	/* Tell our caller what the name of the file and bail. */
 	if (write(STDOUT_FILENO, filename,
-		  strlen(filename)) != strlen(filename)) {
+		  strlen(filename)) != (ssize_t) strlen(filename)) {
 		return 8;
 	}
 	if (isatty(STDOUT_FILENO)) {
